+++ lib/salah/http/bad_response_error.rb
@@ -1,0 +1,7 @@
+class Salah::HTTP::BadResponseError < RuntimeError
+  attr_reader :response
+  def initialize(*args, response)
+    super(*args)
+    @response = response
+  end
+end